requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
groq
pydantic
pdfminer.six
python-docx
